requirements.txt
jupyterlab==2.2.6
matplotlib==3.3.1
numpy==1.18.5
pandas==1.1.1
Pillow==7.2.0
sagemaker==2.5.3
tensorflow-datasets==3.2.1
tensorflow-gpu==2.3.0
